requirements.txt
# Hollywood Studio - Python Dependencies

# Core Dependencies
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# AI/LLM
anthropic>=0.25.0
openai>=1.0.0
groq>=0.4.0

# Audio Processing
pydub>=0.25.1
edge-tts>=6.1.0

# Image Processing
opencv-python>=4.8.0

# Web/API
websocket-client>=1.6.0
urllib3>=2.0.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# GUI (PyQt6)
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0

# Optional: Local Whisper for subtitles
# openai-whisper>=20231117

# Optional: Advanced color grading
# colour-science>=0.4.0
